Set more WiFi powersaving off options.
The rtl8192cu driver is current for Raspbian Bullseye out-the-box, but
does not support the modprobe options. Those are left in in case of
reverting to the proprietary 8192cu driver. For rtl8192cu (and rtl8xxxu
in the future) we can use the various options for
/etc/network/interfaces and iw (unclear which one should be used, so
doing both for good measure).